requirements.txt
Flask
Flask-CORS
opencv-python
requests
Pillow
numpy
flask-socketio
python-socketio[client]